+++ src/index.browser.js
@@ -1,0 +1,19 @@
+import LocationBadge          from './react/location.badge'
+import { LocationBadge$ }     from './react/location.badge'
+
+import LocationButton         from './react/location.button'
+import { LocationButton$ }    from './react/location.button'
+
+import LocationObserver       from './react/location.observer'
+import { LocationObserver$ }  from './react/location.observer'
+
+export {
+  LocationButton,
+  LocationButton$,
+
+  LocationBadge,
+  LocationBadge$,
+
+  LocationObserver,
+  LocationObserver$,
+}